Check if FAUST annotation has as given level for a given marker

Check if FAUST annotation has as given level for a given marker

Usage

.is_faust_ann_a_match_for_marker(faust_ann, marker, level)

Arguments

faust_ann

data.frame. FAUST annotation table read from <project_path>/faustData/sampleData//faustAnnotation.csv.

marker

character. Name of marker.

level

character. Level of marker, e.g. "1", "2" or "3".